Fixed problem with set only working on 32bit of keywords (see ticket:251) Modified house_e implicit enduses to use set instead of string.
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@601 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ core/class.c
@@ -754,7 +754,7 @@ int class_define_map(CLASS *oclass, /**< the object class */
 			else if (proptype==PT_KEYWORD && prop->ptype==PT_set)
 			{
 				char *keyword = va_arg(arg,char*);
-				unsigned long keyvalue = va_arg(arg, int); 
+				unsigned long keyvalue = va_arg(arg, int64); 
 				if (!class_define_set_member(oclass,prop->name,keyword,keyvalue))
 				{
 					errno = EINVAL;
